Checkpoint current status: everything currently compiles except aim/copyinout.c, powerpc/elf_machdep.c, and powerpc/in_cksum.c. Hopefully, I will see some hello world boot messages soon.
Changes are:

- Create a new powerpc/aim64 directory to hold a small number of
  aim64-specific files. Depending on whether we need to have a
  sys/powerpc64 in the end, this could be moved there.
- Move aim/mmu_oea64.c to aim64
- Pacify the compiler about changed types, update trap_subr.S, etc.
- Remove device options for devices that don't exist on any PPC64
  machines.

+++ sys/powerpc/powerpc/bcopy.c
@@ -86,14 +86,14 @@ memcpy(void *dst0, const void *src0, size_t length)
 		/*
 		 * Copy forward.
 		 */
-		t = (int)src;	/* only need low bits */
+		t = (size_t)src;	/* only need low bits */
 
-		if ((t | (int)dst) & wmask) {
+		if ((t | (uintptr_t)dst) & wmask) {
 			/*
 			 * Try to align operands.  This cannot be done
 			 * unless the low bits match.
 			 */
-			if ((t ^ (int)dst) & wmask || length < wsize) {
+			if ((t ^ (uintptr_t)dst) & wmask || length < wsize) {
 				t = length;
 			} else {
 				t = wsize - (t & wmask);
@@ -118,10 +118,10 @@ memcpy(void *dst0, const void *src0, size_t length)
 		 */
 		src += length;
 		dst += length;
-		t = (int)src;
+		t = (uintptr_t)src;
 
-		if ((t | (int)dst) & wmask) {
-			if ((t ^ (int)dst) & wmask || length <= wsize) {
+		if ((t | (uintptr_t)dst) & wmask) {
+			if ((t ^ (uintptr_t)dst) & wmask || length <= wsize) {
 				t = length;
 			} else {
 				t &= wmask;
